Move Location as a top level concern of Spoom
