+++ lib/spoom/deadcode.rb
@@ -5,12 +5,12 @@ require "erubi"
 require "prism"
 
 require_relative "visitor"
+require_relative "location"
 
 require_relative "deadcode/erb"
 require_relative "deadcode/index"
 require_relative "deadcode/indexer"
 
-require_relative "deadcode/location"
 require_relative "deadcode/definition"
 require_relative "deadcode/reference"
 require_relative "deadcode/send"
